+++ example/lib/main.dart
@@ -11,6 +11,7 @@ import 'case/native_view_demo.dart';
 import 'case/platform_view_perf.dart';
 import 'case/popUntil.dart';
 import 'case/return_data.dart';
+import 'case/rotation_transition.dart';
 import 'case/simple_webview_demo.dart';
 import 'case/selection_screen.dart';
 import 'case/state_restoration.dart';
@@ -161,7 +162,8 @@ class _MyAppState extends State<MyApp> {
     'interceptor': (settings, uniqueId) {
       return PageRouteBuilder<dynamic>(
           settings: settings,
-          pageBuilder: (_, __, ___) => const ImagePickerPage(title: "interceptor"));
+          pageBuilder: (_, __, ___) =>
+              const ImagePickerPage(title: "interceptor"));
     },
     'firstFirst': (settings, uniqueId) {
       return PageRouteBuilder<dynamic>(
@@ -170,12 +172,14 @@ class _MyAppState extends State<MyApp> {
     },
     'willPop': (settings, uniqueId) {
       return PageRouteBuilder<dynamic>(
-          settings: settings, pageBuilder: (_, __, ___) => const WillPopRoute());
+          settings: settings,
+          pageBuilder: (_, __, ___) => const WillPopRoute());
     },
     'counter': (settings, uniqueId) {
       return PageRouteBuilder<dynamic>(
           settings: settings,
-          pageBuilder: (_, __, ___) => const CounterPage(title: "Counter Demo"));
+          pageBuilder: (_, __, ___) =>
+              const CounterPage(title: "Counter Demo"));
     },
     'dualScreen': (settings, uniqueId) {
       return PageRouteBuilder<dynamic>(
@@ -306,6 +310,10 @@ class _MyAppState extends State<MyApp> {
       return PageRouteBuilder<dynamic>(
           settings: settings,
           pageBuilder: (_, __, ___) => StateRestorationDemo());
+    },
+    'rotation_transition': (settings, uniqueId) {
+      return PageRouteBuilder<dynamic>(
+          settings: settings, pageBuilder: (_, __, ___) => RotationTranDemo());
     },
     'bottom_navigation': (settings, uniqueId) {
       return PageRouteBuilder<dynamic>(