+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 ## NEXT
 
+## 4.1.1
+1. [Android] 修复FlutterFragment可能会盖住NativeFragment的Bug (#1736)
+2. Update README_CN.md (#1731)
+
+## 4.1.0
+1. [Android]fix popUntil not working (#1718)
+2. Add demo for afterimage test
+3. [Android] 与纯Flutter应用生命周期行为保持一致：应用切换到后台时，暂停帧调度，解决动画残影的问题。[注意] 应用「务必」确保前后台通知事件（onBackground/onForeground）的准确性（必要时通过dispatchBackForegroundEvent接口进行接管），否则可能出现页面假死现象；
+4. [Android]对子类开放didFragmentShow和didFragmentHide (#1726)
+
+## 4.0.4
+1. 修复 onPostPush 和 onPostPush 类型强转失败 (#1707)
+2. [Android] 提前暴露混淆导致的异常，并给出明确指引
+
 ## 4.0.3
 1. 修复`addEventListener`函数中可能的类型转换错误